+++ packages/core/src/interfaces/enums.ts
@@ -114,3 +114,21 @@ export enum Skeletons {
 	PIE = "pie",
 	DONUT = "donut"
 }
+
+/**
+ * enum of all possible attributes used to aling text horizontally
+ */
+export enum TextAnchor {
+	START = "start",
+	MIDDLE = "middle",
+	END = "end"
+}
+
+/**
+ * enum of all possible attributes used to aling text vertically
+ */
+export enum DominantBaseline {
+	BASELINE = "baseline",
+	MIDDLE = "middle",
+	HANGING = "hanging"
+}