+++ packages/react/src/index.js
@@ -6,6 +6,7 @@ import DonutChart from "./donut-chart";
 import LineChart from "./line-chart";
 import PieChart from "./pie-chart";
 import ScatterChart from "./scatter-chart";
+import RadarChart from "./radar-chart";
 
 export {
 	GroupedBarChart,
@@ -15,5 +16,6 @@ export {
 	DonutChart,
 	LineChart,
 	PieChart,
-	ScatterChart
+	ScatterChart,
+	RadarChart
 };