+++ packages/react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.30.21](https://github.com/IBM/carbon-charts/compare/v0.30.20...v0.30.21) (2020-05-11)
+
+**Note:** Version bump only for package @carbon/charts-react
+
+
+
+
+
+## [0.30.20](https://github.com/IBM/carbon-charts/compare/v0.30.19...v0.30.20) (2020-05-08)
+
+**Note:** Version bump only for package @carbon/charts-react
+
+
+
+
+
 ## [0.30.19](https://github.com/IBM/carbon-charts/compare/v0.30.18...v0.30.19) (2020-05-01)
 
 **Note:** Version bump only for package @carbon/charts-react